Added the INDEX attribute which gives the original frame index
