+++ PynPoint/ProcessingModules/StackingAndSubsampling.py
@@ -60,6 +60,8 @@ class StackAndSubsetModule(ProcessingModule):
         if self.m_stacking is None and self.m_random is None:
             return
 
+        non_static = self.m_image_in_port.get_all_non_static_attributes()
+
         im_shape = self.m_image_in_port.get_shape()
         nimages = im_shape[0]
 
@@ -120,6 +122,10 @@ class StackAndSubsetModule(ProcessingModule):
         self.m_image_out_port.set_all(im_new, keep_attributes=True)
         self.m_image_out_port.copy_attributes_from_input_port(self.m_image_in_port)
         self.m_image_out_port.add_attribute("PARANG", parang_new, static=False)
+        if "NFRAMES" in non_static:
+            self.m_image_out_port.del_attribute("NFRAMES")
+        if "INDEX" in non_static:
+            self.m_image_out_port.del_attribute("INDEX")
         self.m_image_out_port.add_history_information("Stack and subset",
                                                       "stacking ="+str(self.m_stacking)+
                                                       "random ="+str(self.m_random))
@@ -165,6 +171,8 @@ class MeanCubeModule(ProcessingModule):
         if self.m_image_in_port.tag == self.m_image_out_port.tag:
             raise ValueError("Input and output port should have a different tag.")
 
+        non_static = self.m_image_in_port.get_all_non_static_attributes()
+
         nframes = self.m_image_in_port.get_attribute("NFRAMES")
 
         self.m_image_out_port.del_all_data()
@@ -187,6 +195,10 @@ class MeanCubeModule(ProcessingModule):
         sys.stdout.flush()
 
         self.m_image_out_port.copy_attributes_from_input_port(self.m_image_in_port)
+        if "NFRAMES" in non_static:
+            self.m_image_out_port.del_attribute("NFRAMES")
+        if "INDEX" in non_static:
+            self.m_image_out_port.del_attribute("INDEX")
         self.m_image_out_port.close_database()
 
 
@@ -271,4 +283,135 @@ class DerotateAndStackModule(ProcessingModule):
 
         self.m_image_out_port.set_all(stack)
         self.m_image_out_port.copy_attributes_from_input_port(self.m_image_in_port)
+        self.m_image_out_port.close_database()
+
+
+class CombineTagsModule(ProcessingModule):
+    """
+    Module for combining tags from multiple database entries into a single tag.
+    """
+
+    def __init__(self,
+                 image_in_tags,
+                 check_attr=True,
+                 name_in="combine_tags",
+                 image_out_tag="im_arr_combined"):
+        """
+        Constructor of CombineTagsModule.
+
+        :param image_in_tags: Tags of the database entries that are read as input and combined.
+        :type image_in_tags: tuple, str
+        :param check_attr: Compare non-static attributes between the tags or combine all non-static
+                           attributes into the new database tag.
+        :type check_attr: bool
+        :param name_in: Unique name of the module instance.
+        :type name_in: str
+        :param image_out_tag: Tag of the database entry that is written as output. Should not be
+                              present in *image_in_tags*.
+        :type image_out_tag: str
+
+        :return: None
+        """
+
+        super(CombineTagsModule, self).__init__(name_in=name_in)
+
+        self.m_image_out_port = self.add_output_port(image_out_tag)
+
+        if image_out_tag in image_in_tags:
+            raise ValueError("The name of image_out_tag can not be present in image_in_tags.")
+
+        self.m_image_in_tags = image_in_tags
+        self.m_check_attr = check_attr
+
+    def run(self):
+        """
+        Run method of the module. Combines the frames of multiple tags into a single output tag
+        and adds the static and non-static attributes. The values of the attributes are compared
+        between the input tags to make sure that the input tags decent from the same data set.
+
+        :return: None
+        """
+
+        self.m_image_out_port.del_all_data()
+        self.m_image_out_port.del_all_attributes()
+
+        if len(self.m_image_in_tags) < 2:
+            raise ValueError("The tuple of image_in_tags should contain at least two tags.")
+
+        memory = self._m_config_port.get_attribute("MEMORY")
+
+        for i, item in enumerate(self.m_image_in_tags):
+            progress(i, len(self.m_image_in_tags), "Running CombineTagsModule...")
+
+            image_in_port = self.add_input_port(item)
+            nimages = image_in_port.get_shape()[0]
+
+            if memory == 0 or memory >= nimages:
+                frames = [0, nimages]
+
+            else:
+                frames = np.linspace(0,
+                                     nimages-nimages%memory,
+                                     int(float(nimages)/float(memory))+1,
+                                     endpoint=True,
+                                     dtype=np.int)
+
+                if nimages%memory > 0:
+                    frames = np.append(frames, nimages)
+
+            for j, _ in enumerate(frames[:-1]):
+                im_tmp = image_in_port[frames[j]:frames[j+1], ]
+                self.m_image_out_port.append(im_tmp)
+
+            static_attr = image_in_port.get_all_static_attributes()
+            non_static_attr = image_in_port.get_all_non_static_attributes()
+
+            for key in static_attr:
+                status = self.m_image_out_port.check_static_attribute(key, static_attr[key])
+
+                if status == 1:
+                    self.m_image_out_port.add_attribute(key, static_attr[key], static=True)
+
+                elif status == -1 and key[0:7] != "History":
+                    warnings.warn('The static keyword %s is already used but with a different '
+                                  'value. It is advisable to only combine tags that descend from '
+                                  'the same data set.' % key)
+
+            for key in non_static_attr:
+                values = image_in_port.get_attribute(key)
+                status = self.m_image_out_port.check_non_static_attribute(key, values)
+
+                if self.m_check_attr:
+                    if key == "PARANG" or key == "STAR_POSITION" or key == "INDEX":
+                        if status == 1:
+                            self.m_image_out_port.add_attribute(key, values, static=False)
+                        else:
+                            for j in values:
+                                self.m_image_out_port.append_attribute_data(key, j)
+
+                    elif key == "NFRAMES":
+                        continue
+
+                    else:
+                        if status == 1:
+                            self.m_image_out_port.add_attribute(key, values, static=False)
+
+                        if status == -1:
+                            warnings.warn('The non-static keyword %s is already used but with '
+                                          'different values. It is advisable to only combine tags '
+                                          'that descend from the same data set.' % key)
+
+                else:
+                    if status == 1:
+                        self.m_image_out_port.add_attribute(key, values, static=False)
+                    else:
+                        for j in values:
+                            self.m_image_out_port.append_attribute_data(key, j)
+
+        sys.stdout.write("Running CombineTagsModule... [DONE]\n")
+        sys.stdout.flush()
+
+        self.m_image_out_port.add_history_information("Database entries combined",
+                                                      str(np.size(self.m_image_in_tags)))
+
         self.m_image_out_port.close_database()
