+++ PynPoint/ProcessingModules/__init__.py
@@ -5,11 +5,11 @@ from BadPixelCleaning import BadPixelSigmaFilterModule, BadPixelInterpolationMod
                              BadPixelMapModule, BadPixelRefinementModule
 from DarkAndFlatCalibration import DarkCalibrationModule, FlatCalibrationModule
 from PSFSubtractionPCA import PSFSubtractionModule, FastPCAModule
-from StackingAndSubsampling import StackAndSubsetModule, MeanCubeModule, DerotateAndStackModule
+from StackingAndSubsampling import StackAndSubsetModule, MeanCubeModule, DerotateAndStackModule, \
+                                   CombineTagsModule
 from StarAlignment import StarAlignmentModule, StarExtractionModule, StarCenteringModule, \
                           ShiftForCenteringModule
-from ImageResizing import CropImagesModule, ScaleImagesModule, AddLinesModule, RemoveLinesModule, \
-                          CombineTagsModule
+from ImageResizing import CropImagesModule, ScaleImagesModule, AddLinesModule, RemoveLinesModule
 from PSFpreparation import PSFpreparationModule, AngleCalculationModule, SortParangModule
 from TimeDenoising import CwtWaveletConfiguration, DwtWaveletConfiguration, \
                           WaveletTimeDenoisingModule, TimeNormalizationModule
